[3966] Rename Term -> Schools::Term
